feat : 약관 컴포넌트 분리
