+++ src/components/Signup/Nickname.tsx
@@ -4,11 +4,11 @@ import { useForm } from 'react-hook-form';
 import { useLocation, useNavigate } from 'react-router-dom';
 import styled, { css } from 'styled-components';
 
-import { IcAfterCheckbox, IcBeforeCheckbox } from '../../asset/icon';
 import { postSignupInfo } from '../../lib/api/signup';
 import { useGetUsernameCheck } from '../../lib/hooks/useGetUsernameCheck';
 import Error404 from '../../pages/Error404';
 import { AddAccountInfo, NicknameInfo } from '../../types/signup';
+import Terms from './Terms';
 
 const Nickname = () => {
   const location = useLocation();
@@ -23,11 +23,6 @@ const Nickname = () => {
   const [isNicknameExists, setIsNicknameExists] = useState<boolean>();
   const [nickname, setNickname] = useState<string>('');
   const [errorMsg, setErrorMsg] = useState<string>();
-  const termAddress = [
-    'https://trusted-fir-e0c.notion.site/8040a51be7c74c7babf71d4ae344e162',
-    'https://trusted-fir-e0c.notion.site/9df42e8f5c7246adb74027814a5c0cc9',
-  ];
-
   const {
     register,
     formState: { errors },
@@ -57,20 +52,6 @@ const Nickname = () => {
     }
   };
 
-  const handleCheck = (e: React.MouseEvent<HTMLElement>, idx?: number) => {
-    const target = e.target as HTMLInputElement;
-    if (target.name === 'all') {
-      isChecked[0] ? setIsChecked([false, false, false]) : setIsChecked([true, true, true]);
-    } else {
-      if (idx) {
-        isChecked[idx] = !isChecked[idx];
-        isChecked[0] = isChecked[1] && isChecked[2] ? true : false;
-
-        setIsChecked([...isChecked]);
-      }
-    }
-  };
-
   const handleSignup = () => {
     postSignupInfo({ email, password }, nickname).then((res) => {
       if (res?.success) {
@@ -83,8 +64,9 @@ const Nickname = () => {
 
   const handleSpace = (e: React.ChangeEvent<HTMLInputElement>) => {
     const currentInputValue = e.target.value;
+    console.log(currentInputValue);
     if (currentInputValue.includes(' ')) {
-      e.target.value = currentInputValue.trim();
+      e.target.value = currentInputValue.replace(' ', '');
     }
   };
 
@@ -117,47 +99,7 @@ const Nickname = () => {
           </StNicknameWrapper>
           <StInputDesc isDuplicate>{errorMsg}</StInputDesc>
 
-          <StTermWrapper>
-            <StAllCheckWrapper>
-              <StCheckboxBtn type="button" name="all" onClick={handleCheck}>
-                {isChecked[0] ? <IcAfterCheckbox /> : <IcBeforeCheckbox />}
-              </StCheckboxBtn>
-
-              <StTermContent>
-                <p>전체 동의</p>
-              </StTermContent>
-            </StAllCheckWrapper>
-            <StDetailTermWrapper>
-              <StDetailTerm>
-                <StCheckboxBtn type="button" name="first" onClick={(e) => handleCheck(e, 1)}>
-                  {isChecked[1] ? <IcAfterCheckbox /> : <IcBeforeCheckbox />}
-                </StCheckboxBtn>
-                <StTermContent>
-                  <span>(필수) </span>
-                  <span> 만 14세 이상이에요 </span>
-                </StTermContent>
-              </StDetailTerm>
-
-              <StDetailTerm>
-                <StCheckboxBtn type="button" name="first" onClick={(e) => handleCheck(e, 2)}>
-                  {isChecked[2] ? <IcAfterCheckbox /> : <IcBeforeCheckbox />}
-                </StCheckboxBtn>
-                <StTermContent>
-                  <span>(필수) </span>
-                  <span>
-                    <p onClick={() => window.open(termAddress[0], '_blank')}>
-                      <u>이용약관</u>
-                    </p>
-                    및
-                    <p onClick={() => window.open(termAddress[1], '_blank')}>
-                      <u>개인정보수집이용</u>
-                    </p>
-                    동의
-                  </span>
-                </StTermContent>
-              </StDetailTerm>
-            </StDetailTermWrapper>
-          </StTermWrapper>
+          <Terms isChecked={isChecked} setIsChecked={() => setIsChecked(isChecked)} />
 
           <StSubmitBtn disabled={isDuplicate || errors.username || isChecked !== Array(3).fill(true) ? true : false}>
             계정 만들기
@@ -249,96 +191,6 @@ const StCheckDuplicationBtn = styled.button<{ disabled: boolean }>`
           background-color: ${({ theme }) => theme.colors.Pic_Color_Gray_Black};
         `}
 `;
-
-const StTermWrapper = styled.article`
-  display: flex;
-  flex-direction: column;
-
-  max-width: 100%;
-  margin-top: 18.2rem;
-`;
-
-const StAllCheckWrapper = styled.section`
-  display: flex;
-
-  max-width: 100%;
-  height: 3.2rem;
-
-  border-left-width: 0rem;
-  border-right-width: 0rem;
-  border-top-width: 0rem;
-  border-bottom: 0.1rem solid ${({ theme }) => theme.colors.Pic_Color_Gray_4};
-`;
-
-const StCheckboxBtn = styled.button`
-  height: 2.2rem;
-  padding: 0;
-  margin: 0;
-
-  border: none;
-  background-color: transparent;
-
-  svg {
-    pointer-events: none;
-  }
-`;
-
-const StTermContent = styled.div`
-  display: flex;
-  align-items: center;
-
-  width: 100%;
-  height: 2.2rem;
-  margin-left: 2.14%;
-
-  color: ${({ theme }) => theme.colors.Pic_Color_Gray_Black};
-  ${({ theme }) => theme.fonts.Pic_Body1_Pretendard_Medium_16};
-
-  p {
-    ${({ theme }) => theme.fonts.Pic_Body1_Pretendard_Medium_16};
-    color: ${({ theme }) => theme.colors.Pic_Color_Gray_Black};
-  }
-
-  span {
-    display: flex;
-
-    gap: 0.5rem;
-    ${({ theme }) => theme.fonts.Pic_Body1_Pretendard_Medium_16};
-  }
-
-  span > p {
-    ${({ theme }) => theme.fonts.Pic_Body1_Pretendard_Medium_16};
-    color: ${({ theme }) => theme.colors.Pic_Color_Gray_3};
-  }
-
-  span > p > u {
-    ${({ theme }) => theme.fonts.Pic_Body1_Pretendard_Medium_16};
-  }
-
-  span:first-child {
-    margin-right: 0.71rem;
-    ${({ theme }) => theme.fonts.Pic_Body1_Pretendard_Medium_16};
-  }
-
-  span:last-child {
-    color: ${({ theme }) => theme.colors.Pic_Color_Gray_3};
-  }
-`;
-
-const StDetailTermWrapper = styled.section`
-  display: flex;
-  flex-direction: column;
-
-  margin-top: 1.6rem;
-
-  gap: 1.3rem;
-`;
-
-const StDetailTerm = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
 const StSubmitBtn = styled.button<{ disabled: boolean }>`
   width: 100%;
   height: 5.8rem;
